+++ push_swap.h
@@ -39,25 +39,29 @@ void		error_exit(const char *msg);
 /*
 ** Declarations from ps_utils.c
 */
-int			has_values_in_range(char **stack, int size, int min, int max);
+int			has_values_in_range(char **stack, int size, long min, long max);
 int			calculate_cost(int size, int pos);
-int			find_best_value_in_range(t_stacks *stacks, int start, int end);
-void		move_to_top(t_stacks *stacks, int value);
+long		find_best_value_in_range(t_stacks *stacks, long start, long end);
+void		move_to_top(t_stacks *stacks, long value);
+
 /*
 ** Declarations from ps_sort_utils.c
 */
 long		find_max_value(char **stack);
 long		find_min_value(char **stack);
-int			find_position(char **stack, int size, int value, int mode);
-int			*get_sorted_values(t_stacks *stacks, int total_size);
+int			find_position(char **stack, int size, long value, int mode);
+long		*get_sorted_values(t_stacks *stacks, int total_size);
+
 /*
 ** Declarations from ps_simple_sort.c
 */
 void		simple_sort(t_stacks *stacks);
+
 /*
 ** Declarations from ps_turk_sort.c
 */
 void		turk_sort(t_stacks *stacks);
+
 /*
 ** Declarations from ps_moves.c
 */
@@ -72,18 +76,21 @@ int			push(char ***stack1, char ***stack2, int size_1, int size_2);
 void		sa(t_stacks *stacks, int fd);
 void		sb(t_stacks *stacks, int fd);
 void		ss(t_stacks *stacks, int fd);
+
 /*
 ** Declarations from ps_rotates.c
 */
 void		ra(t_stacks *stacks, int fd);
 void		rb(t_stacks *stacks, int fd);
 void		rr(t_stacks *stacks, int fd);
+
 /*
 ** Declarations from ps_rotates_reverse.c
 */
 void		rra(t_stacks *stacks, int fd);
 void		rrb(t_stacks *stacks, int fd);
 void		rrr(t_stacks *stacks, int fd);
+
 /*
 ** Declarations from ps_pushes.c
 */
